bug fixed and optimized

+++ src/main/java/com/healthtechbd/backend/dto/AmbulanceProviderDTO.java
@@ -11,7 +11,7 @@ public class AmbulanceProviderDTO extends UserDTO {
     private String bio;
     private String place;
     private Long balance;
-    private Long rating;
+    private Double rating;
     private Long reviewCount;
 
 }
